// src/entity/tabs/organizationTabs.types.ts
export enum OrganizationTabsNameEnum{
    OrganizationMain = 'OrganizationMain',
    OrganizationTraining = 'OrganizationTraining',
    OrganizationPresentations = 'OrganizationPresentations',
    OrganizationInfoPackages = 'OrganizationInfoPackages',
    OrganizationDocuments = 'OrganizationDocuments',
    OrganizationInstructions = 'OrganizationInstructions',
    OrganizationLeaders = 'OrganizationLeaders',
    OrganizationGroup = 'OrganizationGroup',
}